Fixed a bug where Andromeda changed the enzyme properties when settings up its own enzyme settings. Avoided a possible null pointer.

+++ src/main/java/eu/isas/searchgui/SearchHandler.java
@@ -2328,7 +2328,7 @@ public class SearchHandler {
                     Util.deleteDir(outputTempFolder);
                 }
 
-                if (enableAndromeda) {
+                if (enableAndromeda && AndromedaProcessBuilder.getTempFolderPath() != null) {
                     File andromedaTempFolder = new File(AndromedaProcessBuilder.getTempFolderPath());
                     if (andromedaTempFolder.exists()) {
                         Util.emptyDir(andromedaTempFolder);
